Fixed conversion from 'size_t' to '...', possible loss of data on WIN64.

+++ copasi/MIRIAM/CRDFGraph.cpp
@@ -1,9 +1,9 @@
 // Begin CVS Header
 //   $Source: /Volumes/Home/Users/shoops/cvs/copasi_dev/copasi/MIRIAM/CRDFGraph.cpp,v $
-//   $Revision: 1.42 $
+//   $Revision: 1.42.2.1 $
 //   $Name:  $
 //   $Author: shoops $
-//   $Date: 2010/07/16 19:00:07 $
+//   $Date: 2011/01/12 19:03:21 $
 // End CVS Header
 
 // Copyright (C) 2010 by Pedro Mendes, Virginia Tech Intellectual
@@ -533,7 +533,7 @@ bool CRDFGraph::removeEmptyNodes()
   for (; it != end; ++it)
     {
       pNode = it->second;
-      unsigned int Edges =
+      size_t Edges =
         mSubject2Triplet.count(pNode);
 
       if (Edges == 0 ||
