+++ copasi/UI/CQFunctionsWidget.cpp
@@ -1,9 +1,9 @@
 // Begin CVS Header
 //   $Source: /Volumes/Home/Users/shoops/cvs/copasi_dev/copasi/UI/CQFunctionsWidget.cpp,v $
-//   $Revision: 1.10.2.1 $
+//   $Revision: 1.10.2.2 $
 //   $Name:  $
 //   $Author: shoops $
-//   $Date: 2010/11/24 18:07:53 $
+//   $Date: 2011/01/12 19:07:49 $
 // End CVS Header
 
 // Copyright (C) 2010 by Pedro Mendes, Virginia Tech Intellectual
@@ -102,9 +102,9 @@ void CQFunctionsWidget::deleteSelectedFunctions()
 
   for (i = 0; i < imax; i++)
     {
-      if (pSelectionModel->isRowSelected(i, QModelIndex()))
+      if (pSelectionModel->isRowSelected((int) i, QModelIndex()))
         {
-          mappedSelRows.append(mpProxyModel->mapToSource(mpProxyModel->index(i, 0)));
+          mappedSelRows.append(mpProxyModel->mapToSource(mpProxyModel->index((int) i, 0)));
         }
     }
